Removed info about Luis and Maurice; made a functioning navbar with links to actual templates; made most of the home page; setup some of my-work page
