+++ app/templates/about.html
@@ -1,38 +1,4 @@
 {% extends "layout.html" %} 
 {% block content %} 
-    <div class="profile">
-        <div id="profile-picture" class="profile-picture">
-            <img src="./static/img/logo.jpg">
-        </div>
-        <h1>{{ title }}</h1>
-    </div>
-    {% for person in people %}
-        <h2>{{ person.name|title }}</h2>
-        <h3>About Me:</h3>
-        <p>{{person.aboutMe}}</p>
-
-        <h3>Education:</h3>
-        <ul>
-            {% for education in person.education %}
-                <li>
-                    <p><strong>{{education.degree}}</strong> &emsp; &emsp; {{education.completionDate}}</p>
-                    <p>{{education.institution}}</p>
-                </li>
-            {% endfor %}
-        </ul>
-        <br />
-
-        <h3>Work Experience:</h3>
-        <ul>
-            {% for experience in person.workExperience %}
-                <li>
-                    <p><strong>{{experience.role}}</strong> &emsp; &emsp; {{experience.startDate}} - {{experience.endDate}}</p>
-                    <p>{{experience.organization}}</p>
-                    <p>{{experience.roleDescription}}</p>
-                </li>
-            {% endfor %}
-        </ul>
-        <br />
-        <img src={{person.profileImageURL}} style="max-width: 100px; max-height: 100px;"/>
-    {% endfor %} 
+    
 {% endblock %}
