+++ app/templates/contact.html
@@ -1,0 +1,4 @@
+{% extends "layout.html" %} 
+{% block content %} 
+    
+{% endblock %}
