--- app/templates/hobbies.html
@@ -1,30 +1,0 @@
-{% extends "layout.html" %} 
-{% block content %} 
-    <div class="profile">
-        <div id="profile-picture" class="profile-picture">
-            <img src="./static/img/logo.jpg">
-        </div>
-        <h1>{{ title }}</h1>
-    </div>
-    {% for person in people %}
-        <h2>{{ person.name|title }}</h2>
-        <h3>Hobbies:</h3>
-        <ul>
-        {% for hobby in person.hobbies %}
-            <li>
-                <p>
-                    <strong>
-                        {{hobby.name}} :
-                    </strong>
-                </p>
-                {{hobby.description}}
-                <br />
-                {% for pictureURL in hobby.pictureURLs %}
-                    <img src={{pictureURL}} style="max-width: 100px; max-height: 100px;"/>
-                {% endfor %}    
-            </li>
-        {% endfor %}
-        </ul>
-        <br />
-    {% endfor %} 
-{% endblock %}
