+++ app/templates/home.html
@@ -1,0 +1,29 @@
+{% extends "layout.html" %} 
+
+{% block head %}
+    <link href="https://cdn.jsdelivr.net/npm/bootstrap@5.2.0-beta1/dist/css/bootstrap.min.css" rel="stylesheet" integrity="sha384-0evHe/X+R7YkIZDRvuzKMRqM+OrBnVFBL6DOitfPri4tjfHxaWutUpFmBp4vmVor" crossorigin="anonymous">
+{% endblock %}
+
+{% block content %} 
+
+<div class="container col-xxl-8 px-4 pt-5">
+    <div class="row flex-lg-row-reverse align-items-center g-5 py-0">
+        <div class="col-10 col-sm-8 col-lg-6">
+            <img src="./static/img/lucas/lucas.jpg" class="d-block mx-lg-auto img-fluid rounded-circle" alt="Bootstrap Themes" width="700" height="500" loading="lazy">
+        </div>
+        <div class="col-lg-6">
+            <h1 class="display-5 fw-bold lh-1 mb-3">Software and Web Developer</h1>
+            <p class="lead">Passionate about creating software that is useful to people and enables them to achieve what was previously impossible.</p>
+            <div class="d-grid gap-2 d-md-flex justify-content-md-start">
+                <a href="/my-work">
+                    <button type="button" class="btn btn-lg px-4 me-md-2 text-white" style="background-color: #85775f;">See My Work</button>
+                </a>
+                <a href="/contact">
+                    <button type="button" class="btn btn-outline-secondary btn-lg px-4">Get in Touch</button>
+                </a>
+            </div>
+        </div>
+    </div>
+</div>
+
+{% endblock %}
