--- app/templates/index.html
@@ -1,9 +1,0 @@
-{% extends "layout.html" %} 
-{% block content %} 
-    <div class="profile">
-        <div id="profile-picture" class="profile-picture">
-            <img src="./static/img/logo.jpg">
-        </div>
-        <h1>{{ title }}</h1>
-    </div>
-{% endblock %}
