--- app/templates/map.html
@@ -1,13 +1,0 @@
-{% extends "layout.html" %} 
-{% block content %} 
-    <div class="profile">
-        <div id="profile-picture" class="profile-picture">
-            <img src="./static/img/logo.jpg">
-        </div>
-        <h1>{{ title }}</h1>
-        <br/>
-        <br/>
-        <br/>
-        <img src="./static/img/map/map.png" width="500" height="333">
-    </div>
-{% endblock %}
